User and Profile models and viewmodels added. Login, Register and RegisterInfo Views added. Users and Profiles added to dbcontent.

+++ Controllers/LaptopsController.cs
@@ -38,7 +38,7 @@ namespace LaptopStore.Controllers
         }
 
         [Route("Laptops/LaptopDetails")]
-        public ViewResult LaptopDetails(int id)
+        public ViewResult LaptopDetails(long id)
         {
 
             Laptop laptop = _allLaptops.getObjectLaptop(id);
