+++ Data/Models/CartItem.cs
@@ -2,9 +2,9 @@
 {
     public class CartItem
     {
-        public int id { get; set; }
+        public long id { get; set; }
         public Laptop laptop { get; set; }
-        public int price { get; set; }
+        public ulong price { get; set; }
         public string CartId { get; set; }
     }
 }
