+++ Data/Models/Laptop.cs
@@ -2,15 +2,15 @@
 {
     public class Laptop
     {
-        public int id { get; set; }
+        public long id { get; set; }
         public string name { get; set; }
         public string shortDesc { get; set; }
         public string longDesc { get; set; }
-        public ushort price { get; set; } //USD dollar
+        public ulong price { get; set; } //USD dollar
         public string imgUrl { get; set; }
         public bool isFavorite { get; set; }
         public bool isAvailable { get; set; }
-        public int categoryId { get; set; }
+        public long categoryId { get; set; }
         public virtual Category Category { get; set; }
     }
 }
